JBRULES-2835: revert ksession id from long to int

+++ drools-core/src/main/java/org/drools/command/SingleSessionCommandService.java
@@ -19,6 +19,6 @@ package org.drools.command;
 import org.drools.runtime.CommandExecutor;
 
 public interface SingleSessionCommandService extends CommandService {
-	long getSessionId();
+	int getSessionId();
 	void dispose();
 }
